#25 [REFACTOR] auth 관련 pages, components 폴더 구조 동일화
