+++ src/components/auth/nickname/NicknameInput.tsx
